Re-organize GUI packages
`tk.wurst_client.feature.gui` rather than `tk.wurst_client.gui.feature`
