+++ Wurst Client/src/tk/wurst_client/alts/gui/EmailAllowedCharacters.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package tk.wurst_client.gui.alts;
+package tk.wurst_client.alts.gui;
 
 public class EmailAllowedCharacters
 {
