+++ Wurst Client/src/tk/wurst_client/commands/AddAltCmd.java
@@ -14,8 +14,8 @@ import net.minecraft.client.network.NetworkPlayerInfo;
 import net.minecraft.util.StringUtils;
 import tk.wurst_client.Client;
 import tk.wurst_client.alts.Alt;
+import tk.wurst_client.alts.gui.GuiAltList;
 import tk.wurst_client.commands.Cmd.Info;
-import tk.wurst_client.gui.alts.GuiAltList;
 
 @Info(help = "Adds a player or all players on a server to your alt list.",
 	name = "addalt",
