+++ Wurst Client/src/tk/wurst_client/error/gui/report.md
