+++ Wurst Client/src/tk/wurst_client/servers/gui/GuiCleanUp.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package tk.wurst_client.gui.servers;
+package tk.wurst_client.servers.gui;
 
 import java.io.IOException;
 import java.util.ArrayList;
